requirements.txt
openai
genanki
gradio
sqlite3
python-dotenv
